VoiceStudioWinUI/Assets/services/requirements-optimized-universal.txt
# VoiceStudio Ultimate - Optimized Requirements for Voice Cloning
# Universal compatibility optimization for maximum voice cloning performance

# Core PyTorch 2.9 dependencies (optimized for voice cloning)
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0

# Audio processing (optimized versions)
numpy==1.24.3
librosa==0.10.1
scipy==1.11.1
soundfile>=0.12.1,<1.0.0
pydub>=0.25.1,<1.0.0
ffmpeg-python>=0.2.0,<1.0.0

# Voice cloning and TTS (optimized versions)
TTS==0.22.0
transformers==4.30.0
pyannote.audio==4.0.1
whisper==20231117
accelerate>=0.20.0,<1.0.0
datasets>=2.12.0,<3.0.0
tokenizers>=0.13.0,<1.0.0
safetensors>=0.3.0,<1.0.0
huggingface-hub>=0.15.0,<1.0.0

# RVC dependencies (optimized)
numba==0.58.1
llvmlite==0.40.1
fairseq==0.12.2
faiss-cpu==1.7.4
pyworld==0.3.2
resampy>=0.4.2,<1.0.0

# Web framework (optimized)
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart>=0.0.6,<1.0.0
websockets==12.0
aiohttp==3.9.0
aiofiles>=23.0.0,<24.0.0

# Data processing (optimized)
pandas==2.0.0
pydantic==2.0.0
scikit-learn==1.2.2

# Visualization (optimized)
matplotlib==3.7.1
seaborn==0.12.0
plotly>=5.15.0,<6.0.0

# Performance monitoring (optimized)
psutil==5.9.0
py-cpuinfo>=9.0.0,<10.0.0
GPUtil>=1.4.0,<2.0.0
nvidia-ml-py>=11.0.0,<12.0.0

# Security (optimized)
PyJWT>=2.8.0,<3.0.0
cryptography>=3.4.8,<42.0.0

# Development tools (optimized)
pytest>=7.4.0,<8.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0

# Additional utilities (optimized)
rich>=13.0.0,<14.0.0
click>=8.1.0,<9.0.0
tqdm>=4.65.0,<5.0.0
colorama>=0.4.6,<1.0.0
